ai backend/requirements.txt
google-cloud-storage
google-cloud-firestore
google-cloud-aiplatform
tavily-python
pypdf
python-dotenv